34151 Agrupamento de opções de resposta quando multiplas questões complementares
