+++ src/SME.SGP.Dados/Repositorios/RepositorioQuestionario.cs
@@ -2,9 +2,8 @@
 using SME.SGP.Dominio.Entidades;
 using SME.SGP.Dominio.Interfaces;
 using SME.SGP.Infra;
-using System;
 using System.Collections.Generic;
-using System.Text;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace SME.SGP.Dados.Repositorios
@@ -30,7 +29,8 @@ namespace SME.SGP.Dados.Repositorios
                           from questao q 
                           left join opcao_resposta op on op.questao_id = q.id
                           left join opcao_questao_complementar oqc on oqc.opcao_resposta_id = op.id
-                         where q.questionario_id = @questionarioId ";
+                         where q.questionario_id = @questionarioId 
+                        order by q.id, op.id";
 
             var lookup = new Dictionary<long, Questao>();
             await database.Conexao.QueryAsync<Questao, OpcaoResposta, OpcaoQuestaoComplementar, Questao>(query,
@@ -43,14 +43,16 @@ namespace SME.SGP.Dados.Repositorios
                         lookup.Add(q.Id, q);
                     }
 
-                    if (opcaoResposta != null)
+                    var entidadeOpcaoResposta = q.OpcoesRespostas.FirstOrDefault(a => a.Id == opcaoResposta.Id);
+                    if (entidadeOpcaoResposta == null && opcaoResposta != null)
                     {
                         q.OpcoesRespostas.Add(opcaoResposta);
+                        entidadeOpcaoResposta = opcaoResposta;
                     }
 
                     if (OpcaoQuestaoComplementar != null)
                     {
-                        opcaoResposta.QuestoesComplementares.Add(OpcaoQuestaoComplementar);
+                        entidadeOpcaoResposta.QuestoesComplementares.Add(OpcaoQuestaoComplementar);
                     }
 
                     return q;
